+++ src/async/impl/contextEngine.hpp
@@ -44,14 +44,14 @@ namespace async { namespace impl
     public://for task
         static void contextCreate(Coro *coro, size_t stackSize);
         static void contextActivate(Coro *coro);
-        static void contextDeactivate(Coro *coro, std::atomic<size_t> *atomicForSetAfter, size_t valueForSetAfter);
+        static void contextDeactivate(Coro *coro, std::atomic<uint32_t> *atomicForSetAfter, uint32_t valueForSetAfter);
         static void contextDestroy(Coro *coro);
 
     private:
         static __thread Context _rootContext;
         static __thread Context *_currentContext;
-        static __thread std::atomic<size_t> *_atomicForSetAfter;
-        static __thread size_t _valueForSetAfter;
+        static __thread std::atomic<uint32_t> *_atomicForSetAfter;
+        static __thread uint32_t _valueForSetAfter;
 
     private:
 #if PVOID_SIZE == INT_SIZE
